// _FRONTEND_BACKEND_COPY/frontend/src/lib/types.ts
export interface CandleDataPoint {
  date?: string;
  time?: string;
  open: number;
  high: number;
  low: number;
  close: number;
  volume: number;
  isUp?: boolean;
}